requirements.txt
Flask==2.3.3
PyGithub==1.59.1
google-generativeai==0.3.2
requests==2.31.0
matplotlib==3.7.2
plotly==5.16.1
seaborn==0.12.2
Jinja2==3.1.2
Flask-Mail==0.9.1
python-dotenv==1.0.0
Werkzeug==2.3.7
jsonschema==4.19.1
beautifulsoup4==4.12.2
pandas==2.0.3
numpy==1.24.3
Pillow==10.0.0